requirements.txt
Flask==3.0.3
Jinja2==3.1.4
Werkzeug==3.0.3
itsdangerous==2.2.0
Click==8.1.7
torch==2.4.0
transformers==4.44.0
huggingface-hub==0.24.5
numpy==1.26.4
pillow==10.4.0